# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.supportedDispensaryOptions:
    enum:
    - locations
    type: string
    x-enum-varnames:
    - Locations
  github_com_Linkinlog_LeafListr_internal_models.Cannabinoid:
    properties:
      description:
        type: string
      name:
        type: string
      value:
        type: number
    type: object
  github_com_Linkinlog_LeafListr_internal_models.Location:
    properties:
      address:
        type: string
      city:
        type: string
      id:
        type: string
      locationTypes:
        items:
          type: string
        type: array
      name:
        type: string
      state:
        type: string
      zipCode:
        type: string
    type: object
  github_com_Linkinlog_LeafListr_internal_models.Offer:
    properties:
      description:
        type: string
      id:
        type: string
    type: object
  github_com_Linkinlog_LeafListr_internal_models.Price:
    properties:
      discountedTotal:
        type: number
      isDiscounted:
        type: boolean
      total:
        type: number
    type: object
  github_com_Linkinlog_LeafListr_internal_models.Product:
    properties:
      brand:
        type: string
      c:
        items:
          $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Cannabinoid'
        type: array
      ctg:
        type: string
      id:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Price'
      subCtg:
        type: string
      t:
        items:
          $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Terpene'
        type: array
      weight:
        type: string
    type: object
  github_com_Linkinlog_LeafListr_internal_models.Terpene:
    properties:
      description:
        type: string
      name:
        type: string
      value:
        type: number
    type: object
info:
  contact: {}
  description: This is the Listr server for dispensary management.
  title: Listr API
  version: 0.2.0
paths:
  /dispensaries:
    get:
      consumes:
      - application/json
      description: Returns a list of supported dispensaries
      produces:
      - application/json
      responses:
        "200":
          description: List of supported dispensaries" Enums(Curaleaf, Beyond-Hello)
          schema:
            items:
              type: string
            type: array
      summary: List supported dispensaries
      tags:
      - dispensaries
  /dispensaries/{dispensaryId}:
    get:
      consumes:
      - application/json
      description: Returns details of a specific dispensary
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dispensary details
          schema:
            items:
              $ref: '#/definitions/api.supportedDispensaryOptions'
            type: array
      summary: Get dispensary details
      tags:
      - dispensaries
  /dispensaries/{dispensaryId}/locations:
    get:
      consumes:
      - application/json
      description: Returns a list of locations for a specific dispensary
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Menu type
        in: query
        name: menu_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of locations
          schema:
            items:
              $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Location'
            type: array
      summary: List locations for a dispensary
      tags:
      - locations
  /dispensaries/{dispensaryId}/locations/{locationId}:
    get:
      consumes:
      - application/json
      description: Returns details of a specific location
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Location details
          schema:
            items:
              type: string
            type: array
      summary: Get location details
      tags:
      - locations
  /dispensaries/{dispensaryId}/locations/{locationId}/cannabinoids:
    get:
      consumes:
      - application/json
      description: Returns a list of cannabinoids for a specific location
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Menu type
        in: query
        name: menu_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of cannabinoids
          schema:
            items:
              $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Cannabinoid'
            type: array
      summary: List cannabinoids for a location
      tags:
      - cannabinoids
  /dispensaries/{dispensaryId}/locations/{locationId}/categories:
    get:
      consumes:
      - application/json
      description: Returns a list of categories for a specific location
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Menu type
        in: query
        name: menu_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            items:
              type: string
            type: array
      summary: List categories for a location
      tags:
      - categories
  /dispensaries/{dispensaryId}/locations/{locationId}/offers:
    get:
      consumes:
      - application/json
      description: Returns a list of offers for a specific location
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Menu type
        in: query
        name: menu_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of offers
          schema:
            items:
              $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Offer'
            type: array
      summary: List offers for a location
      tags:
      - offers
  /dispensaries/{dispensaryId}/locations/{locationId}/products:
    get:
      consumes:
      - application/json
      description: Returns a list of products for a specific location
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Menu type
        in: query
        name: menu_type
        required: true
        type: string
      - description: Category
        in: query
        name: category
        required: true
        type: string
      - description: Sub Category
        in: query
        name: sub
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Brands to include
        in: query
        name: brands
        type: string
      - description: Brands to exclude
        in: query
        name: not_brands
        type: string
      - description: Variants to include
        in: query
        name: variants
        type: string
      - description: Terms to exclude
        in: query
        name: excludes
        type: string
      - description: Terms to include
        in: query
        name: includes
        type: string
      - description: Sort products
        enum:
        - price_asc
        - price_desc
        in: query
        name: price_sort
        type: string
      - description: Sort products
        enum:
        - thc_asc
        - thc_desc
        in: query
        name: thc_sort
        type: string
      - description: Most important terpene
        in: query
        name: terp1
        type: string
      - description: Second most important terpene
        in: query
        name: terp2
        type: string
      - description: Third most important terpene
        in: query
        name: terp3
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            items:
              $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Product'
            type: array
      summary: List products for a location
      tags:
      - products
  /dispensaries/{dispensaryId}/locations/{locationId}/products/{productId}:
    get:
      consumes:
      - application/json
      description: Returns details of a specific product
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Menu type
        in: query
        name: menu_type
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Product'
      summary: Get product details
      tags:
      - products
  /dispensaries/{dispensaryId}/locations/{locationId}/terpenes:
    get:
      consumes:
      - application/json
      description: Returns a list of terpenes for a specific location
      parameters:
      - description: Dispensary ID
        in: path
        name: dispensaryId
        required: true
        type: string
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      - description: Menu type
        in: query
        name: menu_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of terpenes
          schema:
            items:
              $ref: '#/definitions/github_com_Linkinlog_LeafListr_internal_models.Terpene'
            type: array
      summary: List terpenes for a location
      tags:
      - terpenes
swagger: "2.0"